+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/config/StrucmotifConfig.java
@@ -108,10 +108,6 @@ public class StrucmotifConfig {
      */
     private int commitInterval = 16;
     /**
-     * RCSB URL that reports currently present identifiers in archive.
-     */
-    private String rcsbEntryHoldingsUrl = "https://data.rcsb.org/rest/v1/holdings/current/entry_ids";
-    /**
      * URL of the Chemical Component Dictionary.
      */
     private String ccdUrl = "https://ftp.wwpdb.org/pub/pdb/data/monomers/components.cif.gz";
@@ -503,22 +499,6 @@ public class StrucmotifConfig {
      */
     public void setCommitInterval(int commitInterval) {
         this.commitInterval = commitInterval;
-    }
-
-    /**
-     * Where entry holdings are fetched from.
-     * @return a URL
-     */
-    public String getRcsbEntryHoldingsUrl() {
-        return rcsbEntryHoldingsUrl;
-    }
-
-    /**
-     * Update where holdings data is fetched from.
-     * @param rcsbEntryHoldingsUrl a URL
-     */
-    public void setRcsbEntryHoldingsUrl(String rcsbEntryHoldingsUrl) {
-        this.rcsbEntryHoldingsUrl = rcsbEntryHoldingsUrl;
     }
 
     /**